fake mali api version to be 10.6
there is only one mali blob available for the t720 mali and that
is the one for the allwinner h6. that one requires mali api
version 10.6 - so lets fake that version here too that we will get
at least one step further.

next problem then is that the blob is build for the gpu revision
r1p1 and the gpu in the mt8163seems to be r1p0, so there might be
some binary patching of the blob required to get further ...